*NOT WORKING*  Implemented Review Popup
The code does not work but follows the steps outlined: https://github.com/KjellConnelly/react-native-rate

To run, first run npm install react-native-rate.
When you click on the CardScreen page, it should bring up the review pop-up. If there is no valid address for the review store, it should redirect the user to the crohns website.

The error occurs with the Rate.rate method. The error message says "Cannot read property 'rate' of null"